+++ backend-rust/src/migrations.rs
@@ -17,6 +17,9 @@ mod m0019_payday_stake_information;
 mod m0020_chain_update_events;
 mod m0021_amounts_schedule;
 mod m0023_add_init_parameter;
+mod m0025_fix_passive_delegator_stake;
+mod m0026_update_genesis_validator_info;
+mod m0027_reindex_credential_deployments;
 
 /// Ensure the current database schema version is compatible with the supported
 /// schema version.
@@ -241,6 +244,12 @@ pub enum SchemaVersion {
     AddInputParameterToInitTransactions,
     #[display("0024:Precompute validators APYs")]
     BakerPeriodApyViews,
+    #[display("0025:Fix passive delegators stake")]
+    FixPassiveDelegatorsStake,
+    #[display("0026:Update information for genesis validators")]
+    UpdateGenesisValidatorInfo,
+    #[display("0027:Reindex credential deployments, adjusting cost and missing information")]
+    ReindexCredentialDeployment,
 }
 impl SchemaVersion {
     /// The minimum supported database schema version for the API.
@@ -248,7 +257,7 @@ impl SchemaVersion {
     /// have been introduced since this version.
     pub const API_SUPPORTED_SCHEMA_VERSION: SchemaVersion = SchemaVersion::BakerPeriodApyViews;
     /// The latest known version of the schema.
-    const LATEST: SchemaVersion = SchemaVersion::BakerPeriodApyViews;
+    const LATEST: SchemaVersion = SchemaVersion::ReindexCredentialDeployment;
 
     /// Parse version number into a database schema version.
     /// None if the version is unknown.
@@ -292,6 +301,9 @@ impl SchemaVersion {
             SchemaVersion::FixCorruptedPassiveDelegators => false,
             SchemaVersion::AddInputParameterToInitTransactions => false,
             SchemaVersion::BakerPeriodApyViews => false,
+            SchemaVersion::FixPassiveDelegatorsStake => false,
+            SchemaVersion::UpdateGenesisValidatorInfo => false,
+            SchemaVersion::ReindexCredentialDeployment => false,
         }
     }
 
@@ -326,6 +338,9 @@ impl SchemaVersion {
             SchemaVersion::FixCorruptedPassiveDelegators => false,
             SchemaVersion::AddInputParameterToInitTransactions => false,
             SchemaVersion::BakerPeriodApyViews => false,
+            SchemaVersion::FixPassiveDelegatorsStake => false,
+            SchemaVersion::UpdateGenesisValidatorInfo => false,
+            SchemaVersion::ReindexCredentialDeployment => false,
         }
     }
 
@@ -482,7 +497,16 @@ impl SchemaVersion {
                     .await?;
                 SchemaVersion::BakerPeriodApyViews
             }
-            SchemaVersion::BakerPeriodApyViews => unimplemented!(
+            SchemaVersion::BakerPeriodApyViews => {
+                m0025_fix_passive_delegator_stake::run(&mut tx, endpoints).await?
+            }
+            SchemaVersion::FixPassiveDelegatorsStake => {
+                m0026_update_genesis_validator_info::run(&mut tx, endpoints).await?
+            }
+            SchemaVersion::UpdateGenesisValidatorInfo => {
+                m0027_reindex_credential_deployments::run(&mut tx, endpoints).await?
+            }
+            SchemaVersion::ReindexCredentialDeployment => unimplemented!(
                 "No migration implemented for database schema version {}",
                 self.as_i64()
             ),