+++ frontend/src/components/SuspendedValidators/SuspendedValidatorsContent.vue
@@ -1,25 +1,27 @@
 ﻿<template>
-	<div>
-		<SuspendedValidatorsHeader />
-		<DrawerContent>
-			<SuspendedValidators
-				v-if="
-					data.suspendedValidators?.nodes?.length &&
-					data.suspendedValidators?.nodes?.length > 0
-				"
-				:suspended-validators="data.suspendedValidators!.nodes"
-			/>
-			<div v-else class="p-4">No validators suspended</div>
-			<PrimedForSuspensionValidators
-				v-if="
-					data.primedForSuspensionValidators?.nodes?.length &&
-					data.primedForSuspensionValidators?.nodes?.length > 0
-				"
-				:primed-for-suspension-validators="data.primedForSuspensionValidators!.nodes"
-			/>
-			<div v-else class="p-4">No validators primed for suspension</div>
-		</DrawerContent>
-	</div>
+	<SuspendedValidatorsHeader />
+	<DrawerContent>
+		<SuspendedValidators
+			v-if="
+				data.suspendedValidators?.nodes?.length &&
+				data.suspendedValidators?.nodes?.length > 0
+			"
+			:suspended-validators="data.suspendedValidators!.nodes"
+			:page-info="data.suspendedValidators.pageInfo"
+			:go-to-page="goToPageSuspendedValidators"
+		/>
+		<div v-else class="p-4">No validators suspended</div>
+		<PrimedForSuspensionValidators
+			v-if="
+				data.primedForSuspensionValidators?.nodes?.length &&
+				data.primedForSuspensionValidators?.nodes?.length > 0
+			"
+			:primed-for-suspension-validators="data.primedForSuspensionValidators!.nodes"
+			:page-info="data.primedForSuspensionValidators.pageInfo"
+			:go-to-page="goToPagePrimedForSuspensionValidators"
+		/>
+		<div v-else class="p-4">No validators primed for suspension</div>
+	</DrawerContent>
 </template>
 
 <script lang="ts" setup>
@@ -32,6 +34,8 @@ import type { SuspendedValidatorsType } from '~/queries/useSuspendedValidatorsQu
 
 type Props = {
 	data: SuspendedValidatorsType
+	goToPageSuspendedValidators: PaginationGoToPage
+	goToPagePrimedForSuspensionValidators: PaginationGoToPage
 }
 
 defineProps<Props>()